refactor: add core app with TimeStampedModel and unique_slug
Extracts shared base utilities (TimeStampedModel, unique_slug) into a
new apps.core app. machines/models.py now imports from apps.core.
