+++ backend/apps/core/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CoreConfig(AppConfig):
+    name = "apps.core"
+    verbose_name = "Core"
